Move components connected to store to containers
